+++ internal/tasks/task_single_read.go
@@ -21,7 +21,7 @@ type SingleReadTask struct {
 	Collection      string                  `json:"collection,omitempty" doc:"true"`
 	OperationConfig *SingleOperationConfig  `json:"singleOperationConfig" doc:"true"`
 	Operation       string                  `json:"operation" doc:"false"`
-	ResultSeed      int64                   `json:"resultSeed" doc:"false"`
+	ResultSeed      int                     `json:"resultSeed" doc:"false"`
 	TaskPending     bool                    `json:"taskPending" doc:"false"`
 	result          *task_result.TaskResult `json:"-" doc:"false"`
 	req             *Request                `json:"-" doc:"false"`
@@ -54,7 +54,7 @@ func (task *SingleReadTask) CheckIfPending() bool {
 }
 
 // Config configures  the insert task
-func (task *SingleReadTask) Config(req *Request, seed int64, seedEnd int64, reRun bool) (int64, error) {
+func (task *SingleReadTask) Config(req *Request, seed int, seedEnd int, reRun bool) (int, error) {
 	task.TaskPending = true
 	task.req = req
 
@@ -70,7 +70,7 @@ func (task *SingleReadTask) Config(req *Request, seed int64, seedEnd int64, reRu
 	}
 
 	if !reRun {
-		task.ResultSeed = time.Now().UnixNano()
+		task.ResultSeed = int(time.Now().UnixNano())
 		task.Operation = SingleReadOperation
 		task.result = task_result.ConfigTaskResult(task.Operation, task.ResultSeed)
 
@@ -120,7 +120,7 @@ func (task *SingleReadTask) Do() error {
 
 	singleReadDocuments(task, collection)
 
-	task.result.Success = int64(len(task.OperationConfig.KeyValue)) - task.result.Failure
+	task.result.Success = (len(task.OperationConfig.KeyValue)) - task.result.Failure
 
 	if err := task.result.SaveResultIntoFile(); err != nil {
 		log.Println("not able to save result into ", task.ResultSeed)
